feat: rename and fix odri_ros2_examples
