+++ odri_ros2_examples/config/robots/separate_legs.yaml
