+++ odri_ros2_examples/package.xml
@@ -1,18 +1,20 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>odri_examples</name>
+  <name>odri_ros2_examples</name>
   <version>0.0.0</version>
   <description>TODO: Package description</description>
-  <maintainer email="pepms@todo.todo">pepms</maintainer>
+  <maintainer email="jmarti@iri.upc.edu">pepms</maintainer>
   <license>TODO: License declaration</license>
 
   <buildtool_depend>ament_cmake</buildtool_depend>
 
   <depend>rclcpp</depend>
   <depend>std_msgs</depend>
-  <depend>odri_ros2_msgs</depend>
-  <depend>odri_interface</depend>
+  <depend>odri_ros2_interfaces</depend>
+  <depend>odri_ros2_gazebo</depend>
+  <depend>odri_ros2_hardware</depend>
+  <depend>hidro_ros2_utils</depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>ament_lint_common</test_depend>
